refactor: pages폴더 정리 완료
