--- apps/client/src/pages/ai-results/components/ai-results-empty-view/index.ts
@@ -1,1 +1,0 @@
-export { default as AiResultsEmptyView } from './ai-results-empty-view';
