--- apps/client/src/pages/all-memo/components/all-memo-empty-view/index.ts
@@ -1,1 +1,0 @@
-export { default as AllMemoEmptyView } from './all-memo-empty-view';
