--- apps/client/src/pages/all-memo/components/memo/index.ts
@@ -1,3 +1,0 @@
-export { default as InputContent } from './ui/input-content/input-content';
-export { default as InputTitle } from './ui/input-title/input-title';
-export { default as LabelSelect } from './ui/label-select/label-select';
