--- apps/client/src/pages/login/auth/oauth-redirect-guard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './oauth-redirect-guard';
