--- apps/client/src/pages/login/components/login-callback/index.ts
@@ -1,1 +1,0 @@
-export { default } from './login-callback';
