+++ apps/client/src/pages/login/login-callback-page.tsx
@@ -1,4 +1,4 @@
-import LoginCallback from '@pages/login/components/login-callback';
+import LoginCallback from '@pages/login/components/login-callback/login-callback';
 
 const LoginCallbackPage = () => {
   return <LoginCallback />;
