--- apps/client/src/pages/new-memo/components/index.ts
@@ -1,4 +1,0 @@
-export { default as MemoInput } from './memo-input/memo-input';
-export { default as Tab } from './tab/tab';
-export { default as TabList } from './tab-list/tab-list';
-export { default as ToolBar } from './toolbar/toolbar';
