+++ apps/client/src/shared/features/memo-editor/input-content.tsx
@@ -1,9 +1,9 @@
 import { useMemo } from 'react';
 import ReactQuill from 'react-quill-new';
 
-import '../../libs/quill-register';
+import './libs/quill-register';
 
-import { memoQuillFormats, memoQuillModules } from '../../libs/quill-config';
+import { memoQuillFormats, memoQuillModules } from './libs/quill-config';
 
 import * as styles from './input-content.css';
