+++ apps/client/src/shared/features/memo-editor/label-select.css.ts
