+++ apps/client/src/shared/features/memo-editor/libs/quill-register.ts
